requirements.txt
Django>=3.0.5,<3.1.0
Djangorestframework>=3.11.0,<3.12.0
psycopg2>=2.7.5,<2.8.0

flake8>=3.6.0,<3.7.0

django-countries>=6.1.2,<6.3.0
Pillow>=5.3.0,<5.4.0
stripe>=2.46.0,<2.47.0
